Fixes and app launcher icon changed

+++ app/src/main/java/kz/sayat/weatherapp/ui/WeatherItem.kt
@@ -5,7 +5,7 @@ data class WeatherItem (
         val temperatureKelvin: Float,
         val feelsLikeKelvin: Float,
         val humidity: Int,
-        val windSpeed: Int
+        val windSpeed: Float
 ) {
     val temperatureCelsius: Int
         get() = (temperatureKelvin - 273).toInt()
